move test helpers closer to consumers
